add style hook
