+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "requited",
   "version": "1.0.0",
-  "main": "index.ts",
+  "main": "src/index.ts",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
