+++ src/App.tsx
@@ -1,0 +1,16 @@
+import { StatusBar } from 'expo-status-bar';
+import { Text, View } from 'react-native';
+import useGlobalStyles from './styles/UseGlobalStyles';
+import React from 'react';
+
+export default function App() {
+  const globalStyles = useGlobalStyles()
+
+  return (
+    <View style={globalStyles.container}>
+      <Text >Requited</Text>
+      <StatusBar style="auto" />
+    </View>
+  );
+}
+
