+++ src/styles/UseGlobalStyles.tsx
@@ -1,0 +1,12 @@
+import { StyleSheet } from "react-native"
+
+export default function useGlobalStyles() {
+	return StyleSheet.create({
+		container: {
+			flex: 1,
+			backgroundColor: '#fff',
+			alignItems: 'center',
+			justifyContent: 'center',
+  	},
+	})
+}
